Struct google_cloudtasks2_beta2::api::Queue

pub struct Queue {
    pub app_engine_http_target: Option<AppEngineHttpTarget>,
    pub name: Option<String>,
    pub pull_target: Option<PullTarget>,
    pub purge_time: Option<String>,
    pub rate_limits: Option<RateLimits>,
    pub retry_config: Option<RetryConfig>,
    pub state: Option<String>,
    pub stats: Option<QueueStats>,
    pub task_ttl: Option<String>,
    pub tombstone_ttl: Option<String>,
}

A queue is a container of related tasks. Queues are configured to manage how those tasks are dispatched. Configurable properties include rate limits, retry options, target types, and others.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations queues create projects (request|response)

locations queues get projects (response)

locations queues patch projects (request|response)

locations queues pause projects (response)

locations queues purge projects (response)

locations queues resume projects (response)

Fields
app_engine_http_target: Option<AppEngineHttpTarget>

App Engine HTTP target. An App Engine queue is a queue that has an AppEngineHttpTarget.

name: Option<String>

Caller-specified and required in CreateQueue, after which it becomes output only. The queue name. The queue name must have the following format: projects/PROJECT_ID/locations/LOCATION_ID/queues/QUEUE_ID * PROJECT_ID can contain letters ([A-Za-z]), numbers ([0-9]), hyphens (-), colons (:), or periods (.). For more information, see Identifying projects * LOCATION_ID is the canonical ID for the queue's location. The list of available locations can be obtained by calling ListLocations. For more information, see https://cloud.google.com/about/locations/. * QUEUE_ID can contain letters ([A-Za-z]), numbers ([0-9]), or hyphens (-). The maximum length is 100 characters.

pull_target: Option<PullTarget>

Pull target. A pull queue is a queue that has a PullTarget.

purge_time: Option<String>

Output only. The last time this queue was purged. All tasks that were created before this time were purged. A queue can be purged using PurgeQueue, the App Engine Task Queue SDK, or the Cloud Console. Purge time will be truncated to the nearest microsecond. Purge time will be unset if the queue has never been purged.

rate_limits: Option<RateLimits>

Rate limits for task dispatches. rate_limits and retry_config are related because they both control task attempts however they control how tasks are attempted in different ways: * rate_limits controls the total rate of dispatches from a queue (i.e. all traffic dispatched from the queue, regardless of whether the dispatch is from a first attempt or a retry). * retry_config controls what happens to particular a task after its first attempt fails. That is, retry_config controls task retries (the second attempt, third attempt, etc).

retry_config: Option<RetryConfig>

Settings that determine the retry behavior. * For tasks created using Cloud Tasks: the queue-level retry settings apply to all tasks in the queue that were created using Cloud Tasks. Retry settings cannot be set on individual tasks. * For tasks created using the App Engine SDK: the queue-level retry settings apply to all tasks in the queue which do not have retry settings explicitly set on the task and were created by the App Engine SDK. See App Engine documentation.

state: Option<String>

Output only. The state of the queue. state can only be changed by called PauseQueue, ResumeQueue, or uploading queue.yaml/xml. UpdateQueue cannot be used to change state.

stats: Option<QueueStats>

Output only. The realtime, informational statistics for a queue. In order to receive the statistics the caller should include this field in the FieldMask.

task_ttl: Option<String>

The maximum amount of time that a task will be retained in this queue. Queues created by Cloud Tasks have a default task_ttl of 31 days. After a task has lived for task_ttl, the task will be deleted regardless of whether it was dispatched or not. The task_ttl for queues created via queue.yaml/xml is equal to the maximum duration because there is a storage quota for these queues. To view the maximum valid duration, see the documentation for Duration.

tombstone_ttl: Option<String>

The task tombstone time to live (TTL). After a task is deleted or completed, the task's tombstone is retained for the length of time specified by tombstone_ttl. The tombstone is used by task de-duplication; another task with the same name can't be created until the tombstone has expired. For more information about task de-duplication, see the documentation for CreateTaskRequest. Queues created by Cloud Tasks have a default tombstone_ttl of 1 hour.